backend/requirements.txt
# 后端依赖（FastAPI）
# 部署到 Railway 时使用

fastapi>=0.110.0,<0.120.0
uvicorn[standard]>=0.27.0,<0.30.0
sqlalchemy>=2.0.0,<3.0.0
psycopg[binary]>=3.1.0,<4.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
